fix: Fix option list not showing from inside web component
event.target is undefined inside a web component, because the purpose of
the Shadow DOM is to mask the HTML content the Shadow DOM from the
container point of view. With the first entry of the composedPath array
we still have access to the reference of the element that is triggering
the event, but only as long the Shadow DOM is on open mode.